requirements.txt
openai
requests
pandas
python-dotenv
tiktoken